Quote app
Quoteis a sleek and modern Android  application that to bring a daily dose of inspiration . The app integrates with a Quotes API to seamlessly fetch and share new motivational and thought-provoking quotes.

Tech Stack
Kotlin: Built entirely in Kotlin for concise and expressive code.

Android SDK: Harnessing the power of the Android framework for robust mobile development.

Quotes API: Integration with a reliable API to fetch and display inspirational quotes.